Using the newly moved functions
