+++ src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub mod config;
 pub mod dates_and_times;
 pub mod file_io;
-pub mod misc;
 pub mod work_summary;
